transcriber/pyproject.toml
[tool.poetry]
name = "youtube-transcriber"
version = "0.1.0"
description = "Service to transcribe YouTube audio using Gemini"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{ include = "transcriber", from = "src" }]

[tool.poetry.dependencies]
python = "^3.13"
google-generativeai = "^0.3.1"
google-cloud-storage = "^2.10.0"
pydantic = "^2.0"
python-dotenv = "^1.0.0"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
httpx = "^0.25.0"
pydub = "^0.25.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
